Refactor classes into new Maze::Servlets module
