+++ bin/maze-runner
@@ -18,7 +18,6 @@ require_relative '../lib/maze/helper'
 require_relative '../lib/maze/logger'
 require_relative '../lib/maze/network'
 require_relative '../lib/maze/proxy'
-require_relative '../lib/maze/reflective_servlet'
 require_relative '../lib/maze/retry_handler'
 require_relative '../lib/maze/runner'
 require_relative '../lib/maze/sauce_labs_utils'
@@ -33,6 +32,11 @@ require_relative '../lib/maze/option'
 require_relative '../lib/maze/option/parser'
 require_relative '../lib/maze/option/processor'
 require_relative '../lib/maze/option/validator'
+
+require_relative '../lib/maze/servlets/command_servlet'
+require_relative '../lib/maze/servlets/servlet'
+require_relative '../lib/maze/servlets/log_servlet'
+require_relative '../lib/maze/servlets/reflective_servlet'
 
 require_relative '../lib/maze/assertions/request_set_assertions'
