+++ lib/maze/servlets/command_servlet.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Maze
+  module Servlets
+
+    # Receives log requests sent from the test fixture
+    class CommandServlet < WEBrick::HTTPServlet::AbstractServlet
+
+    end
+  end
+end
